Testes ServicoHorarios

+++ src/main/java/br/edu/ufape/clickconsultas/negocios/servico/ServicoHorarios.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import br.edu.ufape.clickconsultas.dados.InterfaceColecaoHorarios;
 import br.edu.ufape.clickconsultas.negocios.modelo.Horarios;
+import br.edu.ufape.clickconsultas.negocios.servico.exception.ObjetoNaoEncontradoException;
 
 
 @Service
@@ -19,19 +20,29 @@ public class ServicoHorarios implements InterfaceServicoHorarios {
 		return colecaoHorarios.findAll();
 	}
 	
-	public List<Horarios> buscarPorData(LocalDate data) {
-		return colecaoHorarios.findByData(data);
+	public List<Horarios> buscarPorData(LocalDate data) throws ObjetoNaoEncontradoException {
+		List<Horarios> listaH = colecaoHorarios.findByData(data);
+		if(listaH.isEmpty())
+			throw new ObjetoNaoEncontradoException("o", "horario");
+	
+		return listaH;
 	}
 	
-	public Horarios buscarPorId(long id) {
-		return colecaoHorarios.findById(id).orElse(null);
+	public Horarios buscarPorId(long id) throws ObjetoNaoEncontradoException {
+		Horarios horario = colecaoHorarios.findById(id).orElse(null);
+		if(horario == null)
+			throw new ObjetoNaoEncontradoException("o", "horario");
+		
+		return horario;
 	}
 
 	public Horarios salvar(Horarios horario) {
 		return colecaoHorarios.save(horario);		
 	}
 
-	public void remover(long id) {
-		colecaoHorarios.deleteById(id);
+	public void remover(long id) throws ObjetoNaoEncontradoException {
+		Horarios horario = buscarPorId(id);
+		if(horario != null)
+			colecaoHorarios.deleteById(id);
 	}
 }
